[clang][tmo][nfc] Move handleTypedMemoryAttr into SemaTypedMemory.cpp
As part of the TMO robustification work, I needed to make handleTypedMemoryAttr
accessible outside of SemaDeclAttr.cpp. Given that this meant it was no longer
able to be a TU static function, I moved it to SemaTypedMemory.cpp as I've also
been working to cleanly separate out the TMO code.

The robustification also results in changes to this code though, so to make the
review less obnoxious I'm preflighting the code motion on its own.
